split transfer file into several smaller models files
