+++ src/utils/script_args.py
@@ -4,7 +4,7 @@ from datetime import date, timedelta
 from dataclasses import dataclass
 
 from duneapi.api import DuneAPI
-from src.models import AccountingPeriod
+from src.models.accounting_period import AccountingPeriod
 
 
 def previous_tuesday(day: date = date.today()) -> date:
